# docs/swagger.yaml
info:
  contact: {}
paths:
...
  /amrs/info:
    get:
      consumes:
      - application/json
      parameters:
      - description: AMR Serial number
        in: path
        name: serial_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            type: string
      summary: Get a single state
...
swagger: "2.0"